Added Module Thanos for better integration
