+++ src/test/java/es/unizar/webeng/hello/HelloControllerUnitTest.java
@@ -32,31 +32,4 @@ public class HelloControllerUnitTest {
         assertThat(map.get("message"), is(message));
     }
 
-    // Test for thanos app (form)
-    @Test
-    public void testThanos() throws Exception {
-        HashMap<String, Object> map = new HashMap<>();
-        String view = controller.play(null, map);
-        assertThat(view, is("thanos"));
-    }
-
-    // Test for thanos app (response alive)
-    @Test
-    public void testThanos_res_alive() throws Exception {
-        HashMap<String, Object> map = new HashMap<>();
-        String view = controller.play("true", map);
-        assertThat(view, is("thanos_res"));
-        assertThat(map.containsKey("msg2"), is(true));
-        assertThat(map.get("msg2"), is("Congratulations!!"));
-    }
-
-    // Test for thanos app (response dead)
-    @Test
-    public void testThanos_res_dead() throws Exception {
-        HashMap<String, Object> map = new HashMap<>();
-        String view = controller.play("false", map);
-        assertThat(view, is("thanos_res"));
-        assertThat(map.containsKey("msg2"), is(true));
-        assertThat(map.get("msg2"), is("You would dead :("));
-    }
 }
